Fix Add Status Page placeholder
